+++ docs/Changelog.md
@@ -23,6 +23,7 @@
 - Option to Minimize to System Tray on Startup.
 - Option to Minimize to System Tray when Closed.
 - Option to display Mouse Pointer Overlay to make easier to track Mouse Pointer.
+- Refresh retains selected Audio/Video sources/codecs and Webcam
 
 ## v7.0.1